requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlmodel==0.0.16
redis==5.0.1
pydantic-settings==2.2.1
fakeredis==2.21.3
pytest==8.1.1
httpx==0.27.0